Refactor SteadyState render - add windterrain viz
