+++ inductiva/fluids/scenarios/__init__.py
@@ -4,7 +4,6 @@ from .fluid_block import FluidBlock
 from .wind_tunnel import WindTunnel
 from .wind_terrain import Terrain
 from .wind_terrain import WindTerrain
-from .wind_tunnel import WindTunnelOutput
 from .fluid_tank import (
     FluidTank,
     CubicTankOutlet,
