+++ inductiva/templates/wind_terrain/openfoam/files/0/p
@@ -33,7 +33,7 @@ boundaryField
         uniformValue    constant $pressure;
     }
 
-    terrain
+    object
     {
         type            zeroGradient;
     }
